// src/types/employee.types.ts
// src/types/employee.types.ts
export interface EmployeeProfile {
  name: string;
  employee_name: string;
  employee: string;
  user_id: string;
  date_of_birth?: string;
  date_of_joining?: string;
  gender?: string;
  image?: string;
  department?: string;
  designation?: string;
  branch?: string;
  status: string;
}